--- packages/MeComposable/enums.ts
@@ -1,16 +1,0 @@
-/**
- * 缓存实例
- */
-export const storageInstance = {
-  localStorage,
-  sessionStorage
-}
-
-/**
- * 颜色类型
- */
-export enum colorMode {
-  hex = 'hex',
-  hsb = 'hsb',
-  rgb = 'rgb'
-}
