+++ packages/MeComposable/useColorTransform/enum.ts
@@ -1,0 +1,8 @@
+/**
+ * 颜色类型
+ */
+export enum colorMode {
+  hex = 'hex',
+  hsb = 'hsb',
+  rgb = 'rgb'
+}
