+++ packages/MeComposable/useColorTransform/types.ts
@@ -1,0 +1,8 @@
+/**
+ * useColorTransform
+ */
+export namespace USEColorTransform {
+  export interface Option {
+    type: ('rgb' | 'hex' | 'hsb')[]
+  }
+}
