+++ packages/MeComposable/useLocked/types.ts
@@ -1,0 +1,7 @@
+/**
+ * useLocked
+ */
+export namespace USELocked {
+  export type Option = (...args: unknown[]) => Promise<unknown | void>
+  export type ReturnResult = (...args: unknown[]) => Promise<unknown | void>
+}
