+++ packages/MeComposable/useSameTarget/types.ts
@@ -1,0 +1,4 @@
+/**
+ * handleClick
+ */
+export type HandleClick = (e: PointerEvent) => void
