+++ packages/MeComposable/useScroll/types.ts
@@ -1,0 +1,6 @@
+/**
+ * 滚动动画选项
+ */
+export interface ScrollAnimationOption {
+  duration?: number
+}
