+++ packages/MeComposable/useStorage/enums.ts
@@ -1,0 +1,7 @@
+/**
+ * 缓存实例
+ */
+export const storageInstance = {
+  localStorage,
+  sessionStorage
+}
