+++ packages/MeComposable/useStorage/types.ts
@@ -1,0 +1,5 @@
+/**
+ * 缓存类型
+ */
+export type StorageType = 'local' | 'session'
+
