+++ packages/api.ts
@@ -1,4 +1,4 @@
 /**
  * @file API 入口文件 - 此文件没有包含 install 方法，用于按需引入
  */
-export { useValidator, useBind, useVisible, useUtils, useStorage, useLocked, useMoveHandle, useId } from './MeComposable'
+export * from './MeComposable'
