+++ packages/styles/MeColorPicker.less
@@ -95,7 +95,7 @@
         width: calc(100% + @border-width * 2);
         height: calc(100% + @border-width * 2);
         transform: translate(-50%, -50%);
-        border: 1px solid @border-black;
+        border: 1px solid @border-color-default;
         border-radius: @border-radius-circle;
       }
     }
@@ -228,7 +228,7 @@
   padding: @gap;
   background: @bg-default;
   box-shadow: @shadow-box-secondary;
-  
+
   &::before {
     content: '';
     position: absolute;
